Merged PR 12572: Retail Print Module DotNet refactoring
- Migrated all retail device drivers away from DotNet
- Migrated all print methods away from DotNet
- Migrated OPOS cash drawer opening away from DotNet
- Migrated tax free away from DotNet
- Moved all report handling from object output selection now that BC has built-in events for handling reports
- Many smaller refactorings in data join mgt., device drivers, print management codeunits, hardcoded print	 codeunits.
- Cleaned up retail report selection.
- UTF8 encoded mpos print job to allow for m30 printer usage without encoding handling in the xamarin app.
- Various quality of life improvements on retail print templates: Tree view by default on layout & data items, new "if data" iteration type instead of a 1-iteration loop hack, flowfields showing hidden constraints.
- Removed our email report print (handled by official MS extension now)

Unrelated cleanup along the way:
- Deleted bridge control addin (no longer used)
- Deleted transcendence control addin (no longer used)

All the device codeunits are now building their print jobs in the proper device specific encoding right from the start. This eliminates the need for any string encoding conversion hacks and greatly simplifies the entire print stack from device codeunits to local .NET. It fixes what is an almost 10 year old design bug, to use a string to hold raw bytes rather than a bytearray, which wasn't realistic to fix pre-AL because all the windows client clickonce packages had an ancient .dll hardcoded into them which depended on this behavior. Now it is, finally gone..

---

Beware: It has been decided that this PR is being rushed in because a customer is going live soon in MS cloud and NP does not want to push the golive to allow for full testing.
Known issues:

- Lack of upgrade codeunit for report setup in "object output selection" to standard "printer selection".
- Lack of new printnode setup for new BC report events.
- Lack of testing hardware connector integration with new report print events
- Lack of manually testing other device drivers than epson receipts
- Lack of optimizing the allocation in memory in device drivers.
- Lack of test codeunits covering data join mgt. in retail print module.
- Lack of UI for MPOS connector and hardware connector modal pages (need a spinner).
- Lack of hardware connector module import to re-use the same client javascript code as dragonglass instead of duplicating manually.

Everything missing above will continue being worked on in the mean time while the MS cloud customer acts as beta tester.
**If this PR is cherry picked and deployed to version 9 it will break existing customers due to missing upgrade codeunits and testing..**

Related work items: #6465