Merged PR 12744: LineNo check and get in Ecomm. and Magento orders.
Reason why we are changing this is explained in case 465629.

That "overlap" will not happen very often but it will happen sometimes for sure.

So... because it will not happen very often, it means that it would go OK to generate LineNo ourselves, starting from 10000
So I didn't wanted to go for logic find last Line No and increment it for 10000 before each insert because it would be a lot of filtering/SQL queries.

Instead, I have used logic let it work like it works now. Let it to try to insert with Line No starting from 10000 and only when it fail, then filter for last Line No. and check if it exist already. If exist, it means that Insert command failed because of that. Just increase Line No. for 10000 and try insert again.
And if it fail again, that is another issue. Or... if it failed first time because of another issue, we will have that one of course.

Further, why I have created new functions for some checks... I have created because of repeating same code.
And why I haven't created for some other tables (like "Sales Comment Line") is because it exist in one place only.

I have tested this myself in same case container.
First, I have added 3 subscribers OnAfterInsert and added one more line in Sales Line, Puchase Line and Sales Comment Line tables (just to simulate that some more extension will do the same). And then I have tried Magento Sales Order and Return Order and also Ecommerce Sales Order. It all went good.

Then I have removed subscribers and tested the same, just to see if current way of working is OK.

Related work items: #6578