Merged PR 12876: Added script for compiling explicitly with code cops
Added script for compiling explicitly with code cops even when disabled in config.
This is intended to allow a developer workflow to switch from always compiling with code cops to instead only running them explicitly towards the end of a change, as a one-time check.

This is important because code cops are getting slower and slower so by avoiding them on each F5, we increase the iteration speed a lot.

On my laptop time spent locally went from 80 seconds to 5 seconds without analyzers.

Related work items: #6687