Merged PR 12346: Pepper for HWC
This will replace the stargate version of Pepper. The .net codeunits have essentially been replaces with new and old are marked as obsolete.

The old integration name has been renamed to PEPPER_SG to allow all setup to be used for by the new implementation. This way the both versions will be easily maintainable for all pre-cloud versions and customers.

Related work items: #6322