Merged PR 12959: v3 actions should have blocking UI by default
The API was done in the wrong direction. It should be blocking implicitly without opt-out for those few that can be async.

There were also some internal methods that should have been marked as internal to allow for additional parameters such as this, even when WorkflowConfig object is public. This was also fixed here.

Related work items: #6741