Merged PR 13180: Added functionality to Cancel QuickPay when not Captured
This is originally for OttoSuenson, but doesn't make sense to add it to ZAL since this is setup on Payment Gateway.

This is only functionality to support Cancel of quick pay if quick pay is set on Payment Gateway and Cancel Codeunit is set also.

This is tested and can be tested on container: https://os524704.dynamics-retail.net/

--------------
Process description
--------------
On Webshop order is created and payed with Quick Pay. (Payment line & Sales Order in BC is automatically created)
![image (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/13180/attachments/image%20%282%29.png)
On BC Sales Order is opened and deleted. (Cancel is triggered automatically if QuickPay Cancel Codeunit was set in Payment Gateway)
![image (3).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/13180/attachments/image%20%283%29.png)
On QuickPay proces should be registered correctly:
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/13180/attachments/image.png)

Related work items: #6860