Merged PR 13208: Correction to procedure SetBufferValues
Per case 526812 client is receiving an error when clicking on Hide Inactive Varieties. The error was: The length of the string is 22, but it must be less than or equal to 20 characters. Value: MATTE/GLOSS TREK BLACK
Found out through debugging that procedure SetBufferValues in Variety Buffer Table had an incorrect length for  Variety Value. Corrected length to 50 characters.

Related work items: #6887