Merged PR 12847: Prevent same parent and child ID on Magen. Categ.
Case 523708

Prevent same parent and child ID on Magento Category.

This issue showed up in case 523296 on Elding because Styer developed some Elding extension which was subscribed on Item Category table and was creating Magento Category in our "NPR Magento Category" table, but with Parent ID 1 which caused an issue. They (Styer) removed this and switched to normal way of doing (pull Magento Categories from Magento to BC) but just in case, we are going to prevent this for the future.

Related work items: #6633