Merged PR 13388: Add Campaign Discount List Line List page instead of using list part page
When accessing the lines of Campaign Discount from the Item List, the Line Part type of page was used. It caused us that actions weren't visible when this list part page was run as a normal page. In this PR I added a new list page for that purpose with action to reach the Campaign Discount card page.

Related work items: #6990