Merged PR 13414: Cleaned up retention policy code in FR compliance
I am not sure what the purpose of this code is.
It does not belong in FR compliance. On top of that, it also doesn't work correctly in a fresh company because the subscriber to POS session init for double checking setup is inside a [TryFunction] that hits the modify here.

Deleting.

Related work items: #7003