PR 13424: Custom codeunit for running "Post Inventory Cost to G/L" batch job though Job Queue on BC18+
When using standard codeunit 2846 "Post Inventory Cost to G/L" (available on BC18+), we encountered the same issue we had previously on BC17
>The print document can’t be downloaded to a client when it runs in a background task or session. The affected report ID is 1002. Verify that a printer has been configured for the report.

This is somehow related to the fact that the only printers available on the server are the ones NST cannot use  (standard Microsoft's "Print to PDF" and "XPS Document Writer" cannot be used in background sessions).
We've already discovered that the correct solution would be instead of using <Report>.Run() statement the <Report>.SaveAs() approach should be utilized.
Since we cannot change the standard codeunit, and on the other hand we've already had the custom codeunit in place for BC17, now we are going to use the same custom codeunit for BC18 and up.

Case 530039
Related work items: #7010