PR 13461: Possibility to disable payment buttons if there are no lines on the POS sale
System will now allow selecting value 'Auto' for the field "Enabled", when defining POS menu buttons of "Action Type" = 'Payment Type'.
And, as soon as value 'Auto' is selected for "Enabled" field, system will automatically fill in field "Data Source Name" with value 'BUILTIN_SALELINE'.
This will make sure front-end automatically disables the button, if there are no lines on the POS sale.

However, there is a limitation to the functionality: the payment buttons displaying amounts in different currencies must have another value assigned as "Data Source Name" (value 'BUILTIN_SALE'). Thus for those buttons we can have either the amounts displayed, or have the possibility to disable them, if there are no lines in the sale. Unfortunately we cannot have both (if we decided to support both at the same time, this would require refactoring on the front-end level, which is out-of-scope for now).

Case 528791
Related work items: #7032