PR 13739: Refactored DE compliance (Fiskaly integration V2)
We couldn't use Fiskaly V1 for integration anymore (this is the version Stjepan originally built integration upon). We had to refactor to V2.
The most important change is that they obsoleted SDK and introduced an API for communication. With the SDK we needed an Azure function (in order to work with external .dll files provided by Fiskaly SDK), so our solution was built upon working with the Azure function. Now we don’t need the Azure function anymore.

We still missing some of the functions existed previously (there are simply no relevant methods on the new Fiskaly API for that). We decided with Ulla that we were going to start with what we had for now.
We've got a customer on MS Azure cloud. So this is going to be included in version 10 and it will need to be uploaded to the store.

Case 529211
Related work items: #7263