PR 13807: POS Action 'DISCOUNT': fixed 'TotalDiscountAmount' discount type calculation issue
'TotalDiscountAmount' discount type was not applied correctly, when there were multiple lines in the sale, "Prices including VAT" was set to false, and user specified discount amount was set to be including VAT.
The issue was that in cases, when there was a discrepancy due to rounding, system would adjust discount amount on the last sale. And while doing the adjustment system would recalculate discount amount to be excluding VAT for the 2nd time, resulting in incorrect discount amount.

Related work items: #7293