Merged PR 13758: End of Day with Frontend look and feel
The new End of day UI is here.

The current action BALANCE_V3 supports the new look & feel, provided that you configure the end-of-day to use the new UI. That is done on the field `User Experience` on the  `POS End of Day Profile Card`. (Note that this field is hidden and this option is there to enable sneak preview of this UI)

There is a new BALANCE_V4 action that should be used instead. This action is a WF3 action and requires payments to also be version WF3. It uses the frontend to dispatch to the EFT Close workflows and orchestrate the continuation of the End-Of-Day process.

Related work items: #7270