Merged PR 14132: Add scanner import support for different type of scanners
Different types of scanners generate different output in files that needs to be imported into the system. On Inventory Setup, the "Scanner Provider" field was added to define which type of scanner is used.
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/14132/attachments/image.png)

This feature is done using interfaces and enums, meaning when there will be a requirement for a new type of scanner to be added, we can do it easily by extending those.

Related work items: #7460