Merged PR 14046: POS Background Tasks
One of the newer BC versions introduced the concept of Page Background Tasks.

This feature is promising but it's a real shame that Microsoft hardcoded it into the page object type in BC instead of allowing real callback-driven async programming in AL  from codeunits. They could have just called them "Background tasks" instead of "Page background tasks" with almost the same API.

However, we are in luck inside the POS: We have a fullscreen page running all day long and of course it can host these background tasks like any other BC page. It also never switches Rec so they are never automatically cancelled due to user clicking around.

So, the following PR adds a little module that wraps the page background tasks of the POS page allowing you to start them from any POS workflow AL code like this:

![pos_background_task.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/14046/attachments/pos_background_task.png)

It's as the official PBT feature but with an API that is a little bit more intuitive:
- You don't need to open any modal pages on the screen to run async code and you can do it from a codeunit.
- The API forces you to handle errors - because you don't want people to skip error handling :)
- The API forces you to declare a timeout value - because you don't want people to not think about timeouts when writing async code :)
- The completion/error triggers get the same parameter object that the child session got so they can see what the original context was - this prevents you from keeping state elsewhere (such as the DB) making it easier to write stateless completion handlers.
- Instead of using OnRun triggers of codeunits with a magically available Page.GetBackgroundParameters() and Page.SetBackgroundTaskResult() functions everything in this wrapper is instead implemented via an interface+enum combo.

Of course, the downside is, all of this only works inside the POS - And only on call stacks initiated from POS page triggers (meaning any time a workflow calls back to AL which is what happens when you click or scan anything in the POS).

Note: The above feature was not just done for fun, but because we need it right now for our async EFT cloud integrations in BC cloud. In the past years these were using background sessions started via StartSession(); to execute HTTP requests on background sessions while writing their data to DB on records that parent session could poll periodically.
But since **each** **tenant** can only have a maximum of 10 background sessions in BC cloud no matter how many users are purchased, we needed an alternative. Good thing is: **each parent** **session** can have 5 active background task child sessions at a time which is much more charitable. And the return data sent back from child session to parent session makes them much nicer to work with as that means we don't have to abuse the database to share state between threa...