Merged PR 14187: Changing quantity from EAN box
From older versions we have a feature for changing a quantity of sales line from the EAN box, simply entering '\*' character at the end of input. For example, if we want to set the quantity to 5, we would enter '5*' in the EAN box.
However, this didn't work from BC17 because of the bugs in the code. One of the bugs is described more in case 538433. Since some customers already asked for this feature, we created this quick fix which will pollute the same results as the main fix that will be introduced in case 538433.

Related work items: #7489