Merged PR 14435: Only handle event if we can actually determine the ChildRecRef
Only handle event if we can actually determine the `ChildRecRef`

This is a follow up to !14297. If the item provided in the `ParentRecRef` then the NpXml code being executed afterwards would fail. Here we only handle the event if we can actually determine what `ChildRecRef` should be.

Related work items: #7600