Merged PR 14499: CustomOptionLine No retrieval change
CustomOptionLine No retrieval change:

Change BC import so it reads the underscores after it identifies the first # from the item.

This was needed because customer has Magento external_no in the format: "3700677606303_M#00022_430000"

00022 is the Custom Option No and 430000 should be the Custom Option Value Line No.
BC identified the Custom Option No the number with '#' in front and the Line No, the text with the first underscore in front.
So if we had 2 underscores it did not work as expected. The line No was "M#00022_43". At import this caused error: "The value "M#00022_43" can't be evaluated into type Integer."

With this change, the system now identifies as Line No separator the "_" after the "#", so the line no is the correct one

Related work items: #7627