Merged PR 14540: Case system (service tier) callback updates
Case system (service tier) callback updates

Hi @<Mark Stewart Pedersen>, this PR contain two major updates for callback to case system (table "NAV Service Tier User")

**1.** We will not use standard "User" table anymore and it's field "Expiry Date" so this will prevent situation we have had a few weeks ago when BC Online user from Australia couldn't login (even after uninstalling NP Retail extension) because we haven't removed Expiration Date in our case system. Now we have separate table named "NPR Client Diagnostic" and we will use "Expiration Date" from that table. We will still block login so they should contact us. But if they don't want, they can just uninstall NP Retail and login will be possible again.

**2.** Number of WEB requests to our case system will be reduced. Instead of checking for User Locked and User Expired on each login, now it will be different:
When logged in, date and time will be saved in "NPR Client Diagnostic" table. And then, during next hour if he logout or just login in new session:
 - if user wasn't set to expire or wasn't locked checks will not be done for that one hour
 - but if it was locked or expired (and you now unlock it, or remove expiration date) then user will not have to wait for one hour for that change to be applied.

What we are getting here is that in 99% of logins, user is not locked and not expired, so there is no need to check if they are now locked or expired on each login. Latency of one hour is I would say acceptable.

Things we can discuss here is:
1. Is the name of this new table good or should be renamed?
2. Are the field names OK (I am aware that they are not unified, but I kept current names from case system)
3. Do you agree that one hour latency is OK (but as I said, only if it wasn't locked or expired)

Related work items: #7547