Merged PR 14607: Fix media field saving images unnecessary
When using media fields on temporary records, BC will still save the media to the Tenant Media table. This in turn means that we will keep the media but not the record that should reference it.

This PR solves this issue for Magento Pictures by showing the image in a control add-in.

Related work items: #7687