Merged PR 14752: Callback on user login reduce to once per day
Callback to case system on user login reduced to once per day. So from now on, if user is not locked or set to be expired, callback will be done each 24 hours (or on next first login). This means that if user was not set to be expired or locked and you set it in our case system, you will have to wait at least 24 hours for that change.

But if user is set to expire or locked, callback will be done on each login, and this is because of urgent cases when user is not able to login, so he/she will not have to wait 24 hours.

Related work items: #7803