Merged PR 14973: POS Display Profile makeover
The POS Display Profile was wierd to configure since it required the POS Unit No. code and the POS Display Profile code to be the same, in order to work together. This meant that only 1 profile could match together with one POS Unit, ergo not reuseable. This has been changed so you need to select the profile in the POS Unit card page. Also the Screen number and Media downloaded fields are made obsolete in the POS Display Profile table (Display Setup), and have their own table where they can be configured (POS Unit Display). This way device specific information is not part of the generic profile anymore.

Related work items: #7968