Merged PR 14859: Add SFTP to NpXml template and switch to NC Endpoint
NpXml template now has option to send files through SFTP instead FTP only.
Feature is enabled through NC Endpoints and when we already did that, FTP is also transferred to NC

Endpoints and all FTP related fields are obsoleted from Xml Template. And upgrade codeunit is created for creating new or assigning already existing NC Endpoint to template.

A lot of hours were spent on this because of so many changes and everything had to be tested again.
Initially there wasn't error handling through NC Endpoint and No Output was also present.
Now we have errors and outputs.

To bad is that we are going to substitute NpXml templates with Codeunits. :( All of this will be also obsoleted.

Related work items: #7432