Merged PR 15102: Readjusting the Discount calc. on Reatil Journal
Readjusting the Discount calc. on Reatil Journal

Regarding this case:
There is a logic set on “Retail Journal Line” table that is executed when a value is added on the „Item No.“ Field, which populates initial values of the line.
After validating this field a function called „UpdateDiscount“ is executed from the Codeunit „NPR Retail Journal Code“ which is doing the following:
- Looks if there is any Period Discount for this Item and uses that period discount
- If no Period Discount was found it is searching for Mix Discounts and uses that if found
- If any discount was found it sets the appropriate values to fields "Discount Type", "Discount Code", "Quantity for Discount Calc", "Discount Price Incl. Vat" and "Discount Price Excl. VAT", based on found discount

The problem here is that this function is to not covering many scenarios that can happen:
- It is not using the “Discount Priority List” to find the prioritized discount but is always searching the Period Discount first, and Mix Discount second
- It is not looking if the Period Discount is active or not
- There is no support for finding Period Discounts based on “Customer Discount Group” (main problem of this case and why it was created)
- It is always taking the first discount found and not looking to find the best discount if more than one is present

I have added some adjustments to the code so that we can use the most standardized calculation that we have, so now when adding a new line, the calculation should be done properly.

Related work items: #7643