Merged PR 15210: Issue concurrent posting
Changed POS Post Entries behavior

When JQ runs posting there it will aggregate a larger set of records for posting without locking them. And the even it a lock would be applied, there was a commit in the middle that would cleared the locks and allow a concurrent process to grab the same entries and also post them.

The routine is slightly altered by removing the commit and instead running the check code inside a try function (instead of codeunit.run()).

The JQ is also catching errors in the posting processes, allowing a rollback of individual period registers when posting multi period registers.

The sacrifice is the JQ log when there is an unforeseen error (outside of the the RunCheck()) that throws an error. This will not be logged to the posting log.  (This logged was messed up anyway).

The "Gen. Jnl.-Check Line".RunCheck() is always executed (regardless of the StopOnErrorVar setting). When StopOnErrorVar is true, the Error() is thrown. This will typically happen in user mode.

In job queue mode, StopOnError is set to false and if the error is caught by RunCheck(), the posting log should be updated with error message and entries marked with "Error during posting" state.

Related work items: #8195