Merged PR 15475: Hide Lines having Zero Admitted Fix
Fix for problem with the option “Hide Lines having Zero Admitted”.

**Description of problem and how the page was working:**
The page is working with 2 major triggers: OnFindRecord and OnNextRecord.
When we enable the “Hide Lines having Zero Admitted” option the OnFindRecord trigger is activated, and it is doing the following:
- It uses the currently selected record on the lines and checks it if it matches the filter conditions
- If it does not, the page loops again on the same trigger taking now the first record of the list and checks if that record matches the filter conditions
- If even this record does not match, then code execution exits giving us the empty list and newer starts the OnNextRecord.
Here is where the problem lies and therefore when someone selects a record line that would meet the filtering criteria and activates the “Hide Lines having Zero Admitted” option, the page would work as intended as it would normally execute the OnFindRecord trigger, finish it with success, and then go to the OnNextRecord trigger which will find all the remaining records fitting the criteria.
This also applies when we have not selected a record that fits the criteria, but have first record in the list that does, as the OnFindRecord trigger (as mentioned before) loops twice where on the second time it gets the first record which fits the criteria, finishes the OnFindRecord trigger with success and executes OnNextRecord trigger like before getting the rest of the record.

**The fix:**
The solution here was to always force the record which meets the criteria into the OnFindRecord trigger so it can be successfully executed which will afterwards trigger the OnNextRecord and create the needed list.

Related work items: #7659