PR 15573: Page "NPR POS Payment Bin Checkpoint": OnOpenPage performance issue
We had long running SQL query telemetry entries for this function. Particularly line 9 of the function, which is:
`Rec.ModifyAll(Type, POSPaymentBinCheckpoint.Type::ZREPORT);`

SQL statement included loading all the table fields, which is odd, as `ModifyAll()` shouldn't read anything.
I guess it was because the action was on the `Rec` variable, and it somehow needed to load all the records to be displayed on the page before running the `ModifyAll()`.
Hope this will solve the issue.
Refactored other parts of the function to minimize data loaded from the server by either using `SetLoadFields()` or omitting looping through records entirely.

Case 552411
Related work items: #8529