Merged PR 15641: Fixed ConfirmTicketReservation to respond nicer when already confirmed
When ConfirmTicketReservation is requested concurrently, there is a race condition for the response status. The "second" confirm would fail the request with "already confirmed" error. And since "first" request might not have produced the response when data was updated, it would also respond fail.

ConfirmTicketReservation response when confirmed should now be idempotent.

The import list will however list it as "Not Imported" with the error message:
"Token %1 has already been confirmed."

The XmlPort code was also given a full cleanup .

Related work items: #8580