Merged PR 15659: Set Obsolete Pending on ALL Task Queue objects
Set Obsolete Pending on ALL Task Queue objects.

After this PR is completed, we should create new release and announce it on the AppSource.
And after that we could start with removing. (But I don't know if we need to wait for next major version or some time to pass...)

Related work items: #7777