Merged PR 15739: Fallback to payment v1 not working for v2 payment with NULL handler
When no payment v2 was found for the payment type, the intention was that it would fallback to v1 automatically.

The fallback to legacy payment flow is now made deliberate for Voucher, Foreign Voucher and Check.
This will buy us some time to do a proper refactoring for the remaining payment v1 workflows.

Related work items: #8654