Merged PR 15771: GB integ. "foreign card detected" error on TaxFree
Global Blue integration, "foreign card detected" error on Tax Free.

Case 527929

Request which is being sent was OK, and response from Global Blue (it has a lot of lines - 960 in my case) was also OK:

![response.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/15771/attachments/response.png)

And loop was doing an Insert 960 times, but always first range pair because of minor typo in the loop. 1 instead of i.

After the change, all lines are there:

![Bug fixed.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/15771/attachments/Bug%20fixed.png)

Related work items: #8669