Merged PR 15850: Move all case system calls to background session
This PR moves all 4 case system requests to the same background session which is started upon login:
- SaaS validation
- Invoicing/stripe setting checker
- POS Unit/POS Store count updater
- Client Diagnostics with more info for each user.

Zooming out, they could all be in one call, but cleaning it up is beyond the scope of this PR and matters much less now that they all happen in sequence from the same background session.

For BC Cloud there is a limit on the total # of background sessions set to 10. If the call to StartSession ever starts failing then we skip - because telemetry like this does not matter under a lot of load.

The end result is the following:
1. POS will work in BC sandboxes, on developer machines and in partner onprem without our azure key vault being available.
2. The rolecenter will load more quickly since there are no blocking webservice requests happening on the user session.
3. The POS can no longer be blocked via a call home to case system - Instead, blocking will have to be done on another level if really required. For example by disabling their users on our tenants manually or shutting down the tenant manually. (Hosting has hundreds of ways to block a non-paying customer.)

Note: Stripe calls is not affected as it will still call to stripe if a company is marked as using stripe invoicing.
And thus, anyone without azure keyvault access configured to use stripe will still fail (There should be any of these as stripe is intended for prod BC cloud where azure keyvault always works).

Related work items: #8734