Merged PR 15879: Removed check if account has changed for deposits
Removed check if account has changed for deposits.

This will post + and - on the same account - with different texts. These kind of posting were suppressed before. But now that there can be only posting of balancing lines in a set of lines, the "posting buffer" will be empty. And the source lines will remain being marked as unposted.

Related work items: #8766