Merged PR 15848: Avoid loading data in Data Log when checking for new records to process
This PR consists of two commits: 1) removing all the old C/SIDE comments from the module and 2) possibly improving the performance a little bit when checking for new records.

The performance improvement is made under the assumption that a lot of the time we don't actually have any new records to process, and we therefore access some data that does not need to be loaded.

Related work items: #8732