Merged PR 15958: "Client Path" fld removed frm NcEndpointFile table
"Client Path" field removed from NcEndpointFile table.

I removed it directly (without using obsolete pending and removed states) because table and codeunit are internal and pages are not extensible.

No, it cannot be removed in one step because of an error AS0002: The field with name 'Client Path' and ID '105' was found in the previous version, but is missing in the current extension. This might break the upgrade of existing installations.

Reverting commit and going on osbsolete pending.

Related work items: #8871