Merged PR 16041: Add Email Scenario on E-mail Template Header
This PR adds Email Scenario on E-mail Template Header. This is to allow for changing the E-mail Account that is sending the template when it gets sent.

If a customer requires a new scenario to be defined that can be added through an enum extension in a PTE.

The functionality can be tested in container attached to case 552769 (Carlsberg restore).

Related work items: #8927