Merged PR 16153: Merge POS view data
Part of the problem that POS frontend developers have is that our view architecture is not designed as a normal web application.

We have a backend that depends on sticky state, remembering the current view, and serving up data to the frontend based on the state that it remembers.

Instead, it can be much simpler: Backend does not remember view, and serves up either the same thing always or whatever it is being asked to in a request by the frontend.
Once that condition is fulfilled then the frontend team can have views that they own, like a normal client-server architecture, which means they can also use normal libraries for managing UI.

I am also removing the event for adding new views as the POS is not really a solution where we just invent views suddenly - that has never happened before and it most likely won't happen.

Note: this PR is only the first step towards a simpler POS view design.
We still need to eventually deprecate the ability of the backend AL code to switch view as it needs to be fully decoupled from the screen currently rendered in the frontend. (Meaning any view switch should be inside the JS workflow code if needed).

But even this small PR makes it possible for Pardip make a fast MPOS payment/sale view switch (without calling an API to be allowed to change GUI locally) and also Sammy to use React Router on his new layout (without calling an API to be allowed to change GUI locally), then we can deal with the remaining AL code later.

Related work items: #9053