Merged PR 16261: Display error message on Download FTP action
When you click on Download FTP action nothing happens because there was some error but not displayed. Like in the case 556290 there was path was not found on the server:

![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/16261/attachments/image.png)

![image (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/16261/attachments/image%20%282%29.png)

It should be shown somewhere so user can now what's wrong. After the change is like this:

![image (3).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/16261/attachments/image%20%283%29.png)

Related work items: #9192