Merged PR 16332: Ftp Encryption Mode enum change frm int. to public
Case 558953

Based on Gustav's need for a BC17 client extension, Ftp Encryption Mode enum changed from internal to public in order to bypass hardcoding a value of Encryption Mode and have it on config table instead.

And because Azure functions are not supporting other encryption types, enum shouldn't be extensible.
So I have changed Extensible to false when BC version is different than 17 with "assumption" that no one extended or will extend it on BC17.

Related work items: #9500