Merged PR 16591: Pepper Additional Parameters on Trx-Level
Pepper has a parameter XmlAdditionalParameters to allow configuration of soft setup. Added support for specifying setup for this feature.

Created a new top-level payment workflow -> EFT_PEPPER_PAYMENT that has this field on the button to allow specific transactional behavior for specific pepper terminals.

The specific case is Pepper terminal 62 (BE) which required addition information to also allow meal-vouchers, travel passes, etc to work with the PSP.
In this case a parameter such as `<VicVoucherTypeValue>CADEAU-PASS</VicVoucherTypeValue>` would enable one of those features on the terminal.

Related work items: #9859