PR 16624: Retail vouchers: avoid references number clash + new publishers
When generating a reference number for a new retail voucher, system didn't check if the same reference number hasn't been already used for an archived voucher. It did check only among active vouchers.
This was an issue, when we tried to integrate retail vouchers with Shopify for Vibholm.
Now system will do the additional check.
\+ added several new publishers to be able to extend the functionality from PTEs

Case 500730
Related work items: #9885