Merged PR 16712: Fixed gaps in blocking POS action design
It only worked for POS menu buttons configured with Type "Action" and manually selected action.
It did not work when configured for those buttons that were shortcuts i.e. type "Payment", "Customer".

We also forgot to mark TEXT_ENTER as non-blocking when we transferred it to workflow v3, meaning that spam-scanning text into the text box could leave to hanging animations when they were skipped.

This PR requires fixes both on frontend (dragonglass) and in AL so I am awaiting a dragonglass PR to be merged before bundling changes from it into this one:
https://dev.azure.com/navipartner/Dragonglass/_git/dragonglass/pullrequest/16711

Related work items: #9953