Merged PR 16800: M2 Mailer, Welcome Email and sync to web
One issue with the M2 member integration is that member are not synchronized to web until member logs in.
This posses some issues for our customers that need to assist their customers with web related issues when doing sales from POS.

To force the "sync" to web, M2-web will (need to) sync member data when asked to send the welcome email.  This was already possible, but this PR makes some small adjustments to make migration to M2-mailing easier.

Most noticeably is the action to apply the communication profile on all members.

Related work items: #10041